Add initial ver of Scene and Camera

+++ src/survivingit/gameobjects/Creature.java
@@ -7,13 +7,13 @@ public abstract class Creature extends GameVisibleObject {
     private int health;
     private float moveSpeed; // Tiles per second
 
-    public Creature(final float x, final float y, final Sprite sprite, final int health, final int moveSpeed) {
+    public Creature(final float x, final float y, final Sprite sprite, final int health, final float moveSpeed) {
 	super(x, y, sprite);
 	this.health = health;
 	this.moveSpeed = moveSpeed;
     }
 
-    public Creature(final Sprite sprite, final int health, final int moveSpeed) {
+    public Creature(final Sprite sprite, final int health, final float moveSpeed) {
 	super(sprite);
 	this.health = health;
 	this.moveSpeed = moveSpeed;
